pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "iblai-app-cli"
version = "0.2.0"
description = "CLI tool to quickly scaffold IBL.ai applications"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "IBL.ai", email = "support@ibl.ai"}
]
keywords = ["iblai", "cli", "nextjs", "ai", "chat", "agent"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
dependencies = [
    "click>=8.0.0",
    "jinja2>=3.0.0",
    "rich>=13.0.0",
    "inquirer>=3.0.0",
    "anthropic>=0.40.0",
    "openai>=1.0.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
    "pyinstaller>=6.0.0",
    "pyyaml>=6.0.0",
]

[project.scripts]
iblai = "iblai.cli:cli"
iblai-app-cli = "iblai.cli:cli"

[project.urls]
Homepage = "https://ibl.ai"
Documentation = "https://docs.ibl.ai"
Repository = "https://github.com/iblai/iblai-app-cli"
"Bug Tracker" = "https://github.com/iblai/iblai-app-cli/issues"

[tool.setuptools.packages.find]
include = ["iblai*"]

[tool.setuptools.package-data]
iblai = ["templates/**/*"]

[tool.black]
line-length = 100
target-version = ["py311", "py312", "py313"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=iblai --cov-report=term-missing"
